[스프링 - 협업 미션] 배럴(김나은) 미션 제출합니다. (#16)
* chore: 설정파일 나누기

* chore: 설정파일 수정

* chore: 설정파일 수정, DataLo
 !test 로 변환

* feat: 추가요금조회 기능 구현

* docs: step4 todos 정리

* feat: Line 관련 extraFare 항목 추가 기능 구현

* feat: 나이관련 요금정책 추가 기능 구현

* feat: 로그인 정보에 따라 요금산출 기능 추가

* fixed: 자바 버전 jwt 의존성 추가

* feat: memberDao 중복값 확인 기능 구현

* feat: MemberService 가입시 이메일 중복 검사 구현

* feat: MemberRequest 나이 유효성 체크

* refactor: 로그인 실패에 대한 예외 추가

* feat: interceptor를 통한 토큰 유효성 검사 로직 추가

* refactor: 기존 테스트에 auth token 적용

* feat: Station 수정 기능 구현 및 api 문서에 따른 예외 추가

* feat: Line crud 동작시 유효성 검증 로직 service layer 추가

* feat: 구간 추가 삭제에 대한 커스텀 예외 추가

* feat: 지하철 지도 전체 반환 기능 구현

* fix: api 수정으로 인한 테스트 수정

* refactor: LineResponse에 Sections 정보를 담도록 변경

* refactor: Station 패키지 컨벤션 맞게 코드 수정

* refactor: Member 유효성 검증 로직 메서드 추출

* refactor: Line 패키지 메서드 명 수정

- dto 정적 팩토리 메서드
- service id 유효성 검사 메서드

* refactor: auth 패키지 내부 config 분리 및 memberDao를 service로 변경

* refactor: PathService 메서드 분리 및 Fare 불필요 생성자 삭제

* refactor: Service @Transactional 어노테이션 추가

* fixed: Line delete 예외 수정

* fixed: Line Color 확인 쿼리 수정

* fixed: Line 수정관련, 삭제관련 버그 수정

* fixed: Sections 삭제 관련 예외 수정

* fixed: Lines 수정 버그 수정

* refactor: prod 설정 및 예외처리 수정, 사용숫자 상수화

* docs: 리뷰내용 및 todos 정리

* docs: 학습로그 데드락 작성

* refactor: 클래스 단위 트랜잭션 제거

* docs: Bean, Component 학습로그 작성

* docs: check todos

* refactor: 요금정책 도메인 로직 interface 로 수정

* refactor: FarePolicy builder 생성자로 수정

* refactor: 간단한 피드백 반영

* refactor: 예외처리 추상화하여 수정

* refactor: Distance 클래스로 분리

* refactor: 도메인 생성자 정리

* refactor: component 빈등록으로 변경, 학습로그 추가

* style: format codes

* refactor: NullObjectPattern 적용

* style: format codes

* docs: 피드백 및 학습로그 정리

* refactor: 불필요한 transactional 제거

* refactor: Distance 관련 클래스 수정

* docs: 카디널리티 학습로그 추가

* style: format codes

* docs: 학습로그 수정

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,40 +1,62 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.exception.deletion.StationCannotDeleteException;
+import wooteco.subway.exception.duplication.StationNameDuplicatedException;
+import wooteco.subway.exception.notfound.StationNotFoundException;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class StationService {
-    private StationDao stationDao;
+
+    private final StationDao stationDao;
 
     public StationService(StationDao stationDao) {
         this.stationDao = stationDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
+        validateDuplicatedName(stationRequest.getName());
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
+    private void validateDuplicatedName(String name) {
+        if (stationDao.findByName(name).isPresent()) {
+            throw new StationNameDuplicatedException();
+        }
+    }
+
     public Station findStationById(Long id) {
-        return stationDao.findById(id);
+        return stationDao.findById(id)
+                .orElseThrow(StationNotFoundException::new);
     }
 
     public List<StationResponse> findAllStationResponses() {
         List<Station> stations = stationDao.findAll();
-
-        return stations.stream()
-                .map(StationResponse::of)
-                .collect(Collectors.toList());
+        return StationResponse.listOf(stations);
     }
 
     public void deleteStationById(Long id) {
+        findStationById(id);
+        validateDeletableStatus(id);
         stationDao.deleteById(id);
+    }
+
+    private void validateDeletableStatus(Long id) {
+        if (stationDao.countRegisteredStations(id) > 0) {
+            throw new StationCannotDeleteException();
+        }
+    }
+
+    public void updateStationById(Long id, StationRequest stationRequest) {
+        findStationById(id);
+        validateDuplicatedName(stationRequest.getName());
+        stationDao.update(stationRequest.getName(), id);
     }
 }
